QMK-HC-13: Move the Nav Layer to Thumb Neutral (#9)
Move the navigation layer to the neutral, or resting, thumb position.
After using my keyboard for a year, I am learning that the nav layer
is one of the most used layers. It's original configuration was not
ergonomic, as it required too much thumb extension and pressure
given the usage.